controllers

+++ backend/Controllers/AuthController.cs
@@ -68,7 +68,7 @@ namespace Backend.Controllers
                 new Claim(ClaimTypes.Email, user.Email),
                 new Claim(ClaimTypes.GivenName, user.FirstName),
                 new Claim(ClaimTypes.Surname, user.LastName),
-                new Claim(ClaimTypes.Role, user.Role.ToString()),
+                new Claim(ClaimTypes.Role, ((int)user.Role).ToString()),
                 new Claim("CompanyId", user.CompanyId.ToString())
             };
